chore: move useNetworkFee for reusability
